refactor(next): extract middleware path matching
