+++ packages/next/src/middleware-dir/utils.ts
@@ -4,12 +4,14 @@ import { GTRuntime } from 'generaltranslation/runtime';
 import { NextURL } from 'next/dist/server/web/next-url';
 import { parseAcceptLanguage } from 'gt-i18n/internal';
 import { normalizePathname, stripTrailingSlashes } from './pathname';
+import {
+  createPathToSharedPathMap,
+  type PathConfig,
+} from './createPathMatcher';
+import { getSharedPath, inDefaultLocalePaths } from './matchPath';
 
-export { normalizePathname };
-
-export type PathConfig = {
-  [key: string]: string | { [key: string]: string };
-};
+export { createPathToSharedPathMap, getSharedPath, normalizePathname };
+export type { PathConfig };
 
 export type ResponseConfig = {
   type: 'next' | 'rewrite' | 'redirect';
@@ -23,28 +25,6 @@ export type ResponseConfig = {
   resetLocaleCookieName: string;
   localeHeaderName: string;
 };
-
-const DYNAMIC_PATH_SEGMENT_PATTERN = '/[^/]+';
-/** Normalizes each segment once without decoding encoded separators. */
-function normalizePathForMatching(pathname: string): string {
-  return pathname.split('/').map(normalizePathname).join('/');
-}
-
-/** Classifies placeholders before decoding static path content. */
-function createPathPattern(pathname: string): string {
-  pathname = stripTrailingSlashes(pathname);
-  if (!/\[([^\]]+)\]/.test(pathname)) {
-    return normalizePathForMatching(pathname);
-  }
-  return pathname
-    .split(/(\[[^\]]+\])/)
-    .map((part, index) =>
-      index % 2
-        ? '[^/]+'
-        : normalizePathForMatching(part).replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-    )
-    .join('');
-}
 
 function applyBasePath(responseUrl: URL, originalUrl: NextURL) {
   const { basePath } = originalUrl;
@@ -185,124 +165,6 @@ export function getLocalizedPath(
   }
 
   return path;
-}
-
-/**
- * Creates a map of localized paths to shared paths using regex patterns
- */
-export function createPathToSharedPathMap(
-  pathConfig: PathConfig,
-  prefixDefaultLocale: boolean,
-  defaultLocale: string
-): {
-  pathToSharedPath: { [key: string]: string };
-  defaultLocalePaths: string[];
-} {
-  return Object.entries(pathConfig).reduce<{
-    pathToSharedPath: { [key: string]: string };
-    defaultLocalePaths: string[];
-  }>(
-    (acc, [sharedPath, localizedPaths]) => {
-      const { pathToSharedPath, defaultLocalePaths } = acc;
-      // Preserve raw templates for parameter substitution and output URLs.
-      pathToSharedPath[createPathPattern(sharedPath)] = sharedPath;
-
-      if (typeof localizedPaths === 'object') {
-        Object.entries(localizedPaths).forEach(([locale, localizedPath]) => {
-          // Convert the localized path to a regex pattern
-          // Replace [param] with [^/]+ to match any non-slash characters
-          const pattern = createPathPattern(localizedPath);
-          pathToSharedPath[`/${locale}${pattern}`] = sharedPath;
-          if (!prefixDefaultLocale && locale === defaultLocale) {
-            pathToSharedPath[pattern] = sharedPath;
-            defaultLocalePaths.push(pattern);
-          }
-        });
-      }
-      return acc;
-    },
-    { pathToSharedPath: {}, defaultLocalePaths: [] }
-  );
-}
-
-/**
- * Gets the shared path from a given pathname, handling both static and dynamic paths
- */
-export function getSharedPath(
-  standardizedPathname: string,
-  pathToSharedPath: { [key: string]: string },
-  pathnameLocale: string | undefined
-): string | undefined {
-  standardizedPathname = normalizePathForMatching(standardizedPathname);
-  const pathnameWithoutTrailingSlash =
-    stripTrailingSlashes(standardizedPathname);
-  // Try exact match first
-  if (pathToSharedPath[pathnameWithoutTrailingSlash]) {
-    return pathToSharedPath[pathnameWithoutTrailingSlash];
-  }
-
-  // Without locale prefix
-  let pathnameWithoutLocale = undefined;
-  // Only remove locale prefix if the locale prefix is valid
-  if (pathnameLocale) {
-    pathnameWithoutLocale = stripTrailingSlashes(
-      standardizedPathname.replace(/^\/[^/]+/, '')
-    );
-    if (pathToSharedPath[pathnameWithoutLocale]) {
-      return pathToSharedPath[pathnameWithoutLocale];
-    }
-  }
-
-  // Try regex pattern match
-  let candidateSharedPath = undefined;
-  for (const [pattern, sharedPath] of Object.entries(pathToSharedPath)) {
-    if (pattern.includes(DYNAMIC_PATH_SEGMENT_PATTERN)) {
-      // Convert the pattern to a strict regex that matches the exact path structure
-      const regex = new RegExp(`^${pattern}$`);
-      // Exact match
-      if (regex.test(pathnameWithoutTrailingSlash)) {
-        return sharedPath;
-      }
-      // Without locale prefix
-      if (
-        !candidateSharedPath &&
-        pathnameLocale &&
-        regex.test(pathnameWithoutLocale as string)
-      ) {
-        candidateSharedPath = sharedPath;
-      }
-    }
-  }
-  return candidateSharedPath;
-}
-
-/**
- * Checks if the pathname is in the default locale paths
- * @param pathname - The pathname to check
- * @param defaultLocalePaths - The default locale paths
- * @returns true if the pathname is in the default locale paths, false otherwise
- */
-
-function inDefaultLocalePaths(
-  pathname: string,
-  defaultLocalePaths: string[]
-): boolean {
-  pathname = stripTrailingSlashes(normalizePathForMatching(pathname));
-  // Try exact match first
-  if (defaultLocalePaths.includes(pathname)) {
-    return true;
-  }
-
-  // Try regex pattern match
-  for (const path of defaultLocalePaths) {
-    if (path.includes(DYNAMIC_PATH_SEGMENT_PATTERN)) {
-      const regex = new RegExp(`^${path}$`);
-      if (regex.test(pathname)) {
-        return true;
-      }
-    }
-  }
-  return false;
 }
 
 /**
